+++ cogito/ghcommitsink.go
@@ -39,8 +39,7 @@ func (sink GitHubCommitStatusSink) Send() error {
 		MaxRetries:   maxRetries,
 		WaitTime:     waitTime,
 		MaxSleepTime: maxSleepTime,
-		// adds some randomness to sleep time to prevent creating a Thundering Herd
-		Jitter: time.Duration(rand.Intn(30)) * time.Second,
+		Jitter:       time.Duration(rand.Intn(30)) * time.Second,
 	}
 	commitStatus := github.NewCommitStatus(target, sink.Request.Source.AccessToken,
 		sink.Request.Source.Owner, sink.Request.Source.Repo, context, sink.Log)