+++ github/commitstatus.go
@@ -109,7 +109,6 @@ func (s CommitStatus) Add(sha, state, targetURL, description string) error {
 	req.Header.Set("Authorization", "token "+s.token)
 	req.Header.Set("Accept", "application/vnd.github.v3+json")
 	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("X-GitHub-Api-Version", "2022-11-28")
 
 	var response httpResponse
 	timeToSleep := 0 * time.Second // 0 seconds
@@ -121,7 +120,10 @@ func (s CommitStatus) Add(sha, state, targetURL, description string) error {
 		if err != nil {
 			return err
 		}
-		timeToSleep = checkForRetry(response, s.target.WaitTime, s.target.MaxSleepTime, s.target.Jitter)
+		timeToSleep, err = checkForRetry(response, s.target.WaitTime, s.target.MaxSleepTime, s.target.Jitter)
+		if err != nil {
+			return fmt.Errorf("internal error: %s", err)
+		}
 		if timeToSleep == 0 {
 			break
 		}
@@ -179,13 +181,27 @@ func httpRequestDo(req *http.Request) (httpResponse, error) {
 	response.oauthInfo = fmt.Sprintf("X-Accepted-Oauth-Scopes: %v, X-Oauth-Scopes: %v",
 		resp.Header.Get("X-Accepted-Oauth-Scopes"), resp.Header.Get("X-Oauth-Scopes"))
 
-	response.rateLimitRemaining, _ = strconv.Atoi(resp.Header.Get("x-ratelimit-remaining"))
-	rateLimitReset, _ := strconv.Atoi(resp.Header.Get("x-ratelimit-reset"))
-	response.rateLimitReset = time.Unix(int64(rateLimitReset), 0)
-
-	// time.RFC1123 is the format of the Github date header
+	if limit := resp.Header.Get("X-RateLimit-Remaining"); limit != "" {
+		v, err := strconv.Atoi(limit)
+		if err != nil {
+			return httpResponse{}, fmt.Errorf("failed to convert X-RateLimit-Remaining header: %s", err)
+		}
+		response.rateLimitRemaining = v
+	}
+	if reset := resp.Header.Get("X-RateLimit-Reset"); reset != "" {
+		v, err := strconv.Atoi(reset)
+		if err != nil {
+			return httpResponse{}, fmt.Errorf("failed to convert X-RateLimit-Reset header: %s", err)
+		}
+		response.rateLimitReset = time.Unix(int64(v), 0)
+	}
+	// time.RFC1123 is the format of the HTTP Date header
 	// example: Date:[ Mon, 02 Jan 2006 15:04:05 MST ]
-	date, _ := time.Parse(time.RFC1123, resp.Header.Get("date"))
+	// https://datatracker.ietf.org/doc/html/rfc2616#section-14.18
+	date, err := time.Parse(time.RFC1123, resp.Header.Get("Date"))
+	if err != nil {
+		return httpResponse{}, fmt.Errorf("failed to parse the date header: %s", err)
+	}
 	response.date = date
 
 	return response, nil
@@ -242,7 +258,7 @@ func min(a, b int) int {
 }
 
 // checkForRetry determines if we should retry the http request and calculates wait time between retries
-func checkForRetry(res httpResponse, waitTime, maxSleepTime, jitter time.Duration) time.Duration {
+func checkForRetry(res httpResponse, waitTime, maxSleepTime, jitter time.Duration) (time.Duration, error) {
 	retrayalbeStatusCodes := []int{
 		http.StatusInternalServerError, // 500
 		http.StatusBadGateway,          // 502
@@ -254,23 +270,22 @@ func checkForRetry(res httpResponse, waitTime, maxSleepTime, jitter time.Duratio
 	// If you exceed the rate limit, the response will have a 403 status and the x-ratelimit-remaining header will be 0
 	// https://docs.github.com/en/rest/overview/resources-in-the-rest-api?apiVersion=2022-11-28#exceeding-the-rate-limit
 	case res.statusCode == http.StatusForbidden && res.rateLimitRemaining == 0:
-		sleepTime := res.rateLimitReset.Sub(res.date) + jitter
-		if sleepTime > maxSleepTime {
-			return 0
-		} else {
-			// Don't allow negative sleep times. Minimum sleep time should be equal to jitter time.
-			return max(sleepTime, jitter)
+		// Calculate sleeptime based solely on the server clock. This is unaffected
+		// by the inevitable clock drift between server and client.
+		sleepTime := res.rateLimitReset.Sub(res.date)
+		// Be a good netizen by adding some jitter to the time we sleep.
+		sleepTime += jitter
+		switch {
+		case sleepTime > maxSleepTime:
+			return 0, nil
+		case sleepTime > 0 && sleepTime < maxSleepTime:
+			return sleepTime, nil
+		default:
+			return 0, fmt.Errorf("unexpected: negative sleep time: %s", sleepTime)
 		}
 	case slices.Contains(retrayalbeStatusCodes, res.statusCode):
-		return waitTime
+		return waitTime, nil
 	default:
-		return 0
+		return 0, nil
 	}
-}
-
-func max(x, y time.Duration) time.Duration {
-	if x < y {
-		return y
-	}
-	return x
 }